testing casting unsigned int

+++ more_singly_linked_lists/7-get_nodeint.c
@@ -8,7 +8,7 @@ listint_t *get_nodeint_at_index(listint_t *head, unsigned int index)
 {
 	unsigned int count = 0;
 
-	while (head != NULL && index != head->n)
+	while (head != NULL && index != (unsigned int)head->n)
 	{
 		count++;
 		head = head->next;
